WHYYYYY DO YOU WORK AND NOT WORKKKKK ITS THE SAME SH*T :japanese_ogre:
